pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "stack_of_luv_times"
version = "0.1.0"
description = "FastAPI application"
dependencies = [
  "fastapi[standard]",
  "requests",
  "python-dotenv",
  "discord",
]